+++ randovania/game_description/resources/resource_database.py
@@ -1,27 +1,35 @@
-from typing import List, NamedTuple, Dict, Optional
+import dataclasses
+from typing import List, Dict, Optional, Callable
 
 from randovania.game_description.resources import search
-from randovania.game_description.resources.damage_resource_info import DamageResourceInfo
+from randovania.game_description.resources.damage_resource_info import DamageReduction
 from randovania.game_description.resources.item_resource_info import ItemResourceInfo
-from randovania.game_description.resources.resource_info import ResourceInfo
+from randovania.game_description.resources.resource_info import ResourceInfo, CurrentResources
 from randovania.game_description.resources.resource_type import ResourceType
 from randovania.game_description.resources.simple_resource_info import SimpleResourceInfo
 from randovania.game_description.resources.trick_resource_info import TrickResourceInfo
 from randovania.games.game import RandovaniaGame
 
 
-class ResourceDatabase(NamedTuple):
+def default_base_damage_reduction(db: "ResourceDatabase", current_resources: CurrentResources):
+    return 1.0
+
+
+@dataclasses.dataclass(frozen=True)
+class ResourceDatabase:
     game_enum: RandovaniaGame
     item: List[ItemResourceInfo]
     event: List[SimpleResourceInfo]
     trick: List[TrickResourceInfo]
-    damage: List[DamageResourceInfo]
+    damage: List[SimpleResourceInfo]
     version: List[SimpleResourceInfo]
     misc: List[SimpleResourceInfo]
     requirement_template: Dict[str, "Requirement"]
+    damage_reductions: Dict[SimpleResourceInfo, List[DamageReduction]]
     energy_tank_item_index: int
     item_percentage_index: Optional[int]
     multiworld_magic_item_index: int
+    base_damage_reduction: Callable[["ResourceDatabase", CurrentResources], float] = default_base_damage_reduction
 
     def get_by_type(self, resource_type: ResourceType) -> List[ResourceInfo]:
         if resource_type == ResourceType.ITEM:
@@ -62,3 +70,12 @@ class ResourceDatabase(NamedTuple):
     @property
     def multiworld_magic_item(self) -> ItemResourceInfo:
         return self.get_item(self.multiworld_magic_item_index)
+
+    def get_damage_reduction(self, resource: SimpleResourceInfo, current_resources: CurrentResources):
+        multiplier = self.base_damage_reduction(self, current_resources)
+
+        for reduction in self.damage_reductions.get(resource, []):
+            if reduction.inventory_item is None or current_resources.get(reduction.inventory_item, 0) > 0:
+                multiplier *= reduction.damage_multiplier
+
+        return multiplier