+++ randovania/layout/preset.py
@@ -1,7 +1,7 @@
 import dataclasses
 import json
 import uuid
-from typing import Optional, List, Iterator, Tuple, Union
+from typing import Optional, List, Iterator, Tuple, Union, Type, Dict
 
 from randovania.bitpacking import bitpacking
 from randovania.bitpacking.bitpacking import BitPackDecoder, BitPackValue
@@ -15,7 +15,8 @@ def _dictionary_byte_hash(data: dict) -> int:
     return bitpacking.single_byte_hash(json.dumps(data, separators=(',', ':')).encode("UTF-8"))
 
 
-_game_to_config = {
+AnyGameConfiguration = Union[PrimeConfiguration, EchoesConfiguration, CorruptionConfiguration]
+_game_to_config: Dict[RandovaniaGame, Type[AnyGameConfiguration]] = {
     RandovaniaGame.PRIME1: PrimeConfiguration,
     RandovaniaGame.PRIME2: EchoesConfiguration,
     RandovaniaGame.PRIME3: CorruptionConfiguration,
@@ -29,7 +30,7 @@ class Preset(BitPackValue):
     description: str
     base_preset_uuid: Optional[uuid.UUID]
     game: RandovaniaGame
-    configuration: Union[PrimeConfiguration, EchoesConfiguration, CorruptionConfiguration]
+    configuration: AnyGameConfiguration
 
     @property
     def as_json(self) -> dict: