+++ randovania/resolver/event_pickup.py
@@ -5,6 +5,7 @@ from randovania.game_description.game_description import GameDescription
 from randovania.game_description.game_patches import GamePatches
 from randovania.game_description.node import EventNode, PickupNode, ResourceNode, Node
 from randovania.game_description.requirements import ResourceRequirement, Requirement
+from randovania.game_description.resources.resource_database import ResourceDatabase
 from randovania.game_description.resources.resource_info import ResourceGain, CurrentResources, ResourceInfo
 
 
@@ -30,15 +31,17 @@ class EventPickupNode(ResourceNode):
     def requirement_to_leave(self, patches: GamePatches, current_resources: CurrentResources) -> Requirement:
         return ResourceRequirement(self.pickup_node.pickup_index, 1, False)
 
-    def can_collect(self, patches, current_resources: CurrentResources, all_nodes: Tuple[Node, ...]) -> bool:
-        event_collect = self.event_node.can_collect(patches, current_resources, all_nodes)
-        pickup_collect = self.pickup_node.can_collect(patches, current_resources, all_nodes)
+    def can_collect(self, patches: GamePatches, current_resources: CurrentResources,
+                    all_nodes: Tuple[Node, ...], database: ResourceDatabase) -> bool:
+        event_collect = self.event_node.can_collect(patches, current_resources, all_nodes, database)
+        pickup_collect = self.pickup_node.can_collect(patches, current_resources, all_nodes, database)
         return event_collect or pickup_collect
 
-    def resource_gain_on_collect(self, patches, current_resources: CurrentResources, all_nodes: Tuple[Node, ...]
+    def resource_gain_on_collect(self, patches, current_resources: CurrentResources,
+                                 all_nodes: Tuple[Node, ...], database: ResourceDatabase,
                                  ) -> ResourceGain:
-        yield from self.event_node.resource_gain_on_collect(patches, current_resources, all_nodes)
-        yield from self.pickup_node.resource_gain_on_collect(patches, current_resources, all_nodes)
+        yield from self.event_node.resource_gain_on_collect(patches, current_resources, all_nodes, database)
+        yield from self.pickup_node.resource_gain_on_collect(patches, current_resources, all_nodes, database)
 
 
 def replace_with_event_pickups(game: GameDescription):