+++ test/resolver/test_resolver_reach.py
@@ -30,7 +30,9 @@ def test_possible_actions_no_resources():
     assert options == []
     prop_a.assert_called_once_with()
     prop_b.assert_called_once_with()
-    node_b.can_collect.assert_called_once_with(state.patches, state.resources, reach._logic.game.world_list.all_nodes)
+    node_b.can_collect.assert_called_once_with(state.patches, state.resources,
+                                               reach._logic.game.world_list.all_nodes,
+                                               state.resource_database)
 
 
 def test_possible_actions_with_event():
@@ -48,6 +50,8 @@ def test_possible_actions_with_event():
     # Assert
     assert options == [event]
     prop.assert_called_once_with()
-    event.can_collect.assert_called_once_with(state.patches, state.resources, reach._logic.game.world_list.all_nodes)
+    event.can_collect.assert_called_once_with(state.patches, state.resources, reach._logic.game.world_list.all_nodes,
+                                              state.resource_database)
     logic.get_additional_requirements.assert_called_once_with(event)
-    logic.get_additional_requirements.return_value.satisfied.assert_called_once_with(state.resources, 1)
+    logic.get_additional_requirements.return_value.satisfied.assert_called_once_with(state.resources, 1,
+                                                                                     state.resource_database)